requirements.txt
flask
python-dotenv
ipykernel
torch
pillow
transformers
openai
opencv-python
langchain-openai
langchain_community
langchain_chroma
pandas
